I did to screenshot 37
This may take a while. Nemo is in pretty much every screenshot (or is in the same space) so I wasn't sure if we wanted to create an endless list of screenshot for this species. This site is pretty neat and can help for ID: https://fishesofaustralia.net.au/home/family/106